change template

+++ src/template.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 
 use crate::challenge::Day;
 
-pub fn day() -> Day<i32> {
+pub fn day() -> Day<usize> {
     Day {
         part1_solutions: (todo!(), None),
         part2_solutions: None,
@@ -13,10 +13,10 @@ pub fn day() -> Day<i32> {
     }
 }
 
-fn part1(data: &str) -> Result<i32> {
+fn part1(data: &str) -> Result<usize> {
     todo!()
 }
 
-fn part2(_data: &str) -> Result<i32> {
+fn part2(_data: &str) -> Result<usize> {
     todo!()
 }
